backend/app/infrastructure/database/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class DBConfig:
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL', 'postgresql://postgres:Henry1587@db:5432/vuelos_db')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SQLALCHEMY_ENGINE_OPTIONS = {
        'pool_size': 5,
        'pool_timeout': 30,
        'pool_recycle': 3600
    }