backend/requirements.txt
Flask==3.1.0
Flask_Caching==2.3.1
flask_restx==1.3.0
flask_sqlalchemy==3.1.1
marshmallow==3.26.1
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
flask-cors==5.0.1
Flask-Migrate==4.1.0
